Take control of virtual hostnames
Before we can sanitise virtual hostnames (see the next commit), we need to
control their values.

Avoid Spring Cloud determining either the virtual hostname or the secure
virtual hostname.

This includes ensuring that both the virtual hostname and the secure virtual
hostname are set for both route and direct registrations. This may not be
strictly necessary, but it is hard to be certain from the Eureka documentation
and code and it should be harmless.

[#115706515]