Sanitise virtual hostnames
When setting secure and unsecure virtual hostnames from an application name,
replace each character which is invalid in a hostname with a dash ('-'). If
any characters were replaced, log a warning message. (Note: if no characters
were replaced, we don't log a warning for virtual hostnames which could
potentially clash with a sanitised virtual hostname, i.e. those which contain
no invalid characters and at least one dash. Perhaps we should.)

Periods ('.') are permitted and will produce a hostname with multiple
components.

Note: RFC 1123 modified the syntax rules for hostnames, but preserved the
character set defined by RFC 952.

It was necessary to subclass EurekaInstanceConfigBean to avoid unpredictable
values of virtual hostname and secure virtual hostname due to setter
injection. These properties are final in the subclass. The corresponding
setters are overridden to throw an exception.

[#115706515]