Provide better defaulting of Eureka app identifiers.
When eureka.instance.appname is defined, default
eureka.instance.virtualHostname and
eureka.instance.secureVirtualHostname to that value.
    
If eureka.instance.appname is not defined but spring.application.name is
defined, default the 3
eureka.instance.appname|virtualHostname|secureVirtualHostname values to
the hostname-sanitized version of spring.application.name.
    
This also addresses
https://github.com/spring-cloud/spring-cloud-netflix/issues/1516 and is
compatible with past and upcoming Camden releases.

[finishes #133115289, #135258295]