Don't suppress Rabbit autoconfig when bound to RMQ
There's some detailed discussion on this in GitHub issue #36.

We have some logic in the connector which tries to disable the standard
Boot RabbitMQ autoconfiguration in cases where it's not needed. Rabbit
always ends up on the classpath due to our transitive dependencies for Hystrix,
but if there's no user-configured RMQ instance to connect to, the
resulting autoconfiguration means we end up with a failing healthcheck.

In a cloud environment, Spring Cloud Connectors provide a Rabbit connection
factory if the app is bound to a Rabbit service instance. But there are other
beans provided by the standard Boot autoconfiguration that an app could still
depend on. These end up missing from the application context because the
conditions for our autoconfig supression logic were still being met.

This commit makes the suppression logic cloud aware. Now we allow the
default Boot Rabbit autoconfig to run if _either_ the user has
configured a Rabbit binding through the environment, _or_ the cloud
environment is providing a Rabbit service binding. This should fix the
missing beans issues.

[finishes #155640825]